point skyline emu to official website
skyline-emu.com is not the official website, it hosts an older build of skyline and while it’s not malicious (I have checked it) it could very well change in the future if more people start visiting this site. I urge you to merge this pr to prevent people from downloading a malicious apk from a fake site. The real website is skyline-emu.one